Minor: Make entire button tile tappable. Change: Don't require Product to have image/emoji in order to appear as a shoppable tile.
